+++ common/src/main/java/it/crystalnest/fancy_entity_renderer/api/entity/player/state/FancyPlayerRenderState.java
@@ -1,13 +1,11 @@
 package it.crystalnest.fancy_entity_renderer.api.entity.player.state;
 
 import it.crystalnest.fancy_entity_renderer.api.Rotation;
-import it.crystalnest.fancy_entity_renderer.api.entity.player.FancyPlayerRenderer;
 import it.crystalnest.fancy_entity_renderer.api.entity.player.FancyPlayerWidget;
 import net.minecraft.client.player.AbstractClientPlayer;
 import net.minecraft.client.renderer.entity.state.AvatarRenderState;
 import net.minecraft.world.entity.Pose;
 import net.minecraft.world.item.ItemStack;
-import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
@@ -17,12 +15,6 @@ import java.util.List;
  * Extension of {@link AvatarRenderState}.
  */
 public class FancyPlayerRenderState extends AvatarRenderState {
-  /**
-   * Internal {@link FancyPlayerRenderer} reference.
-   */
-  @ApiStatus.Internal
-  public FancyPlayerRenderer renderer = null;
-
   /**
    * Left arm rotation.
    */
@@ -148,7 +140,9 @@ public class FancyPlayerRenderState extends AvatarRenderState {
    */
   public boolean showBody = true;
 
-  // TODO: Check (before it was in the superclass). Maybe we can remove this and use only nameTag...
+  /**
+   * Player name.
+   */
   public String name = "Steve";
 
   /**